+++ phpmyfaq/assets/src/utils/index.ts
@@ -5,3 +5,4 @@ export * from './password';
 export * from './reading-time';
 export * from './tooltip';
 export * from './TranslationService';
+export * from './theme-switcher';